:hankey: Combine transect and grid plots into a big subplot figure
Hastily making all-in-one figures for comparing inverted basal parameters over DeepBedMap_DEM and BedMachine Antarctica. The top subplot is a wide A to B transect plot for a given field (e.g. basal_drag, slipperiness, effective pressure, velocity, etc). The bottom panel is from left to right, DeepBedMap, BedMachine, and a diff of the two. Used some low-level `pygmt` magic to do the subplots. The colorbars also needed a bit of work but happy with it now. Also changed to triangulating with float32 instead of float64 as it was giving incorrect number ranges.